refactoring

+++ philo_bonus/sources/dinner.c
@@ -6,7 +6,7 @@
 /*   By: alisa <alisa@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/14 18:53:12 by alisa             #+#    #+#             */
-/*   Updated: 2021/09/15 19:56:42 by alisa            ###   ########.fr       */
+/*   Updated: 2021/09/15 20:07:13 by alisa            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,6 @@ int	take_forks(t_main *m, int philo_name)
 		return (kill(0, SIGTERM));
 	if (sem_wait(m->forks))
 		return (kill(0, SIGTERM));
-	
 	if (sem_wait(m->print))
 			return (kill(0, SIGTERM));
 	if (print_status(m, philo_name, "has taken a fork", NOT_EAT))
@@ -27,7 +26,6 @@ int	take_forks(t_main *m, int philo_name)
 		return (kill(0, SIGTERM));
 	if (sem_wait(m->forks))
 		return (kill(0, SIGTERM));
-	
 	if (sem_wait(m->print))
 			return (kill(0, SIGTERM));
 	if (print_status(m, philo_name, "has taken a fork", NOT_EAT))
@@ -39,15 +37,8 @@ int	take_forks(t_main *m, int philo_name)
 	return (OK);
 }
 
-int	philo_eats(t_main *m, int philo_name)
+static int	check_if_full(t_main *m)
 {
-	if (take_forks(m, philo_name))
-		return (ERROR);
-	if (sem_wait(m->print))
-			return (kill(0, SIGTERM));
-	if (print_status(m, philo_name, "\033[33mis eating\033[0m", EAT))
-		return (ERROR);
-	m->curr_num_of_meals++;
 	if (m->info.num_of_meals != -1 && m->curr_num_of_meals >= m->info.num_of_meals)
 	{
 		if (sem_post(m->print))
@@ -60,6 +51,20 @@ int	philo_eats(t_main *m, int philo_name)
 			return (kill(0, SIGTERM));
 		return (FULL);
 	}
+	return (OK);
+}
+
+int	philo_eats(t_main *m, int philo_name)
+{
+	if (take_forks(m, philo_name))
+		return (ERROR);
+	if (sem_wait(m->print))
+			return (kill(0, SIGTERM));
+	if (print_status(m, philo_name, "\033[33mis eating\033[0m", EAT))
+		return (ERROR);
+	m->curr_num_of_meals++;
+	if (check_if_full(m))
+		return (FULL);
 	if (sem_post(m->print))
 		return (kill(0, SIGTERM));	
 	if (usleep(m->info.time_to_eat))
